core/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pyzmq==26.0.3
openai==1.30.1
pydantic==2.7.2
websockets==12.0
python-dotenv==1.0.1
